Small changes + added home enviroment

+++ src/main/java/com/demoapp/demoapp/entities/StripeItem.java
@@ -33,7 +33,7 @@ public class StripeItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private BigDecimal id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "invoice_id", nullable = false)
